Fix docs on remote development mode
The cli-script example showing how to create an admin account
contains a reload command. Without any parameters this reloads the
server in normal mode. If this script is executed during
configuration of the bootable jar, this causes unexpected side
effects as the server was originally started in embedded mode
(which implies admin-mode) and switches to normal mode after
reloading. This, for instance, results in the application artifact
being deployed and started during server configuration which is
likely to cause errors as configuration is not completed yet.